Add provider map to its own file.
